Move test to correct location and fix assertion
