+++ okapi-logical/src/test/scala/org/opencypher/okapi/logical/impl/logical/LogicalOptimizerTest.scala
@@ -30,9 +30,10 @@ import org.opencypher.okapi.api.schema.Schema
 import org.opencypher.okapi.api.types.{CTNode, _}
 import org.opencypher.okapi.ir.api.expr._
 import org.opencypher.okapi.ir.api.{Label, _}
-import org.opencypher.okapi.testing.MatchHelper._
 import org.opencypher.okapi.logical.impl._
 import org.opencypher.okapi.testing.BaseTestSuite
+import org.opencypher.okapi.testing.MatchHelper._
+import org.opencypher.okapi.trees.BottomUp
 
 import scala.language.implicitConversions
 
@@ -131,6 +132,32 @@ class LogicalOptimizerTest extends BaseTestSuite with IrConstruction {
     )
 
     optimizedLogicalPlan should equalWithTracing(expected)
+  }
+
+  it("should replace cross with value join if filter is present") {
+    val startA = Start(LogicalCatalogGraph(testQualifiedGraphName, testGraphSchema), SolvedQueryModel.empty)
+    val startB = Start(LogicalCatalogGraph(testQualifiedGraphName, testGraphSchema), SolvedQueryModel.empty)
+    val varA = Var("a")(CTNode)
+    val propA = expr.Property(varA, PropertyKey("name"))(CTString)
+    val varB = Var("b")(CTNode)
+    val propB = expr.Property(varB, PropertyKey("name"))(CTString)
+    val equals = Equals(propA, propB)(CTBoolean)
+    val irFieldA = IRField(varA.name)(varA.cypherType)
+    val irFieldB = IRField(varB.name)(varB.cypherType)
+
+    val scanA = NodeScan(varA, startA, SolvedQueryModel(Set(irFieldA)))
+    val scanB = NodeScan(varB, startB, SolvedQueryModel(Set(irFieldB)))
+    val cartesian = CartesianProduct(scanA, scanB, SolvedQueryModel(Set(irFieldA, irFieldB)))
+    val filter = Filter(equals, cartesian, SolvedQueryModel(Set(irFieldA, irFieldB)))
+
+    val optimizedPlan = BottomUp[LogicalOperator](LogicalOptimizer.replaceCartesianWithValueJoin).transform(filter)
+
+    val projectA = Project(propA -> None, scanA, scanA.solved)
+    val projectB = Project(propB -> None, scanB, scanB.solved)
+    val solved = SolvedQueryModel(Set(irFieldA, irFieldB)).withPredicate(equals)
+    val valueJoin = ValueJoin(projectA, projectB, Set(equals), solved)
+
+    optimizedPlan should equalWithTracing(valueJoin)
   }
 
   private def logicalPlan(query: String, schema: Schema): LogicalOperator = {
